refactor(QueuedDrawers): move log function to utils
