+++ apps/ledger-live-mobile/src/newArch/components/QueuedDrawer/index.tsx
@@ -8,7 +8,7 @@ import { isModalLockedSelector } from "~/reducers/appstate";
 import { Merge } from "~/types/helpers";
 import { IsInDrawerProvider } from "~/context/IsInDrawerContext";
 import { DrawerInQueue, useQueuedDrawerContext } from "./QueuedDrawersContext";
-import { logDrawer } from "./QueuedDrawersContextProvider";
+import { logDrawer } from "./utils/logDrawer";
 
 // Purposefully removes isOpen prop so consumers can't use it directly
 export type Props = Merge<
