+++ apps/ledger-live-mobile/src/newArch/components/QueuedDrawer/utils/logDrawer.tsx
@@ -1,0 +1,10 @@
+import { log } from "@ledgerhq/logs";
+import { getEnv } from "@ledgerhq/live-env";
+
+export function logDrawer(
+  message: string | number,
+  data?: Record<string, unknown> | number | string,
+) {
+  if (getEnv("LOG_DRAWERS"))
+    log("QueuedDrawer", typeof message === "number" ? message.toString() : message, data);
+}
